+++ frontend/src/apis/updateHost.ts
@@ -1,15 +1,6 @@
 import { AxiosResponse } from 'axios';
 import { fetchInstance } from '.';
-
-export interface HostInfo {
-  userId: string;
-  liveTitle: string;
-  hostName: string;
-  notice: string;
-  defaultThumbnailImageUrl: string;
-  category: string;
-  tags: string[];
-}
+import { HostInfo } from '@type/hostInfo';
 
 export const updateHost = async (hostInfo: HostInfo): Promise<HostInfo> => {
   const response: AxiosResponse<HostInfo> = await fetchInstance().post('/host/update', hostInfo);
