--- frontend/src/components/host/Form/types.ts
@@ -1,9 +1,0 @@
-export interface FormValues {
-  liveTitle: string;
-  category: string;
-  tag: string;
-  tags: string[];
-  notice: string;
-  hostName: string;
-  previewImage: string | null;
-}
